test(contrib/es): move static test to ESService suite
